+++ src/OrgStat/Logic.hs
@@ -19,7 +19,8 @@ import OrgStat.CLI (CommonArgs (..))
 import OrgStat.Config (ConfOutput (..), ConfOutputType (..), OrgStatConfig (..))
 import OrgStat.Helpers (resolveOutput, resolveReport)
 import OrgStat.Logging (logDebug, logInfo)
-import OrgStat.Outputs (genBlockOutput, genSummaryOutput, processTimeline, tpColorSalt, writeReport)
+import OrgStat.Outputs (genBlockOutput, genSummaryOutput, processScriptOutput, processTimeline,
+                        tpColorSalt, writeReport)
 import OrgStat.WorkMonad (WorkM, wcCommonArgs, wcConfig)
 
 -- | Main application logic.
@@ -29,37 +30,46 @@ runOrgStat = do
     logDebug $ "Config: \n" <> show conf
 
     curTime <- liftIO getZonedTime
-
-    let createDir = do
-            let reportDir = confOutputDir </>
+    let getOutputDir = do
+            let createDir = do
+                    let reportDir =
+                            confOutputDir </>
                             formatTime defaultTimeLocale "%F-%H-%M-%S" curTime
-            liftIO $ createDirectoryIfMissing True reportDir
-            pure reportDir
-    reportDir <- maybe createDir pure =<< views wcCommonArgs caOutputDir
-    logInfo $ "This report set will be put into: " <> fromString reportDir
+                    liftIO $ createDirectoryIfMissing True reportDir
+                    pure reportDir
+            maybe createDir pure =<< views wcCommonArgs caOutputDir
+    let writeOutput coName res = do
+            reportDir <- getOutputDir
+            let prePath = reportDir </> T.unpack coName
+            logInfo $ "This output will be written into: " <> fromString reportDir
+            writeReport prePath res
 
     cliOuts <- views wcCommonArgs caOutputs
     outputsToProcess <-
         bool (mapM resolveOutput cliOuts) (pure confOutputs) (null cliOuts)
+
     forM_ outputsToProcess $ \ConfOutput{..} -> do
         logInfo $ "Processing output " <> coName
-        let prePath = reportDir </> T.unpack coName
         case coType of
             TimelineOutput {..} -> do
                 resolved <- resolveReport toReport
                 let timelineParamsFinal =
                         (confBaseTimelineParams <> toParams) & tpColorSalt .~ confColorSalt
-                let res = processTimeline timelineParamsFinal resolved
-                logInfo $ "Generating timeline report " <> coName <> "..."
-                writeReport prePath res
+                let timeline = processTimeline timelineParamsFinal resolved
+                logDebug $ "Generating timeline report " <> coName <> "..."
+                writeOutput coName timeline
             SummaryOutput params -> do
                 summary <- genSummaryOutput params
-                writeReport prePath summary
+                writeOutput coName summary
+            ScriptOutput params ->
+                processScriptOutput params
             BlockOutput {..} -> do
                 resolved <- resolveReport boReport
                 let res = genBlockOutput boParams resolved
-                writeReport prePath res
+                writeOutput coName res
+
     whenM (views wcCommonArgs caXdgOpen) $ do
+        reportDir <- getOutputDir
         logInfo "Opening reports using xdg-open..."
         void $ shell ("for i in $(ls "<>T.pack reportDir<>"/*); do xdg-open $i; done") empty
     logInfo "Done"