+++ src/OrgStat/Outputs.hs
@@ -6,6 +6,7 @@ module OrgStat.Outputs
 
 import OrgStat.Outputs.Block as Exports
 import OrgStat.Outputs.Class as Exports
+import OrgStat.Outputs.Script as Exports
 import OrgStat.Outputs.Summary as Exports
 import OrgStat.Outputs.Timeline as Exports
 import OrgStat.Outputs.Types as Exports